+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # BOT CHANGELOG
 
+## [v8.7.0] - 2026-01-23
+- `join` - added optional lobby name parameter
+- `string` command - searches followups of a move
+
 ## [v8.6.0] - 2026-01-22
 - fixed DustLoop move notes URL formatting
 - reworded the `fd` error syntax